Add more WellKnownFunctions be more forgiving with arguments
This change adds several methods to WellKnownFunctions that are already being called by falling back to reflection today. For example, the following property function was recently added to the .NET SDK that ends up getting called around 6,000 times via reflection when building Roslyn.slnx:

```
$([System.Text.RegularExpressions.Regex]::IsMatch(%(Identity), '^((\d+\.){3})1$')
```

In addition, there are several WellKnownFunctions that aren't matched during argument checking and fall back to reflection. For example, `String.IsNullOrEmpty(...)` and `String.IsNullOrWhiteSpace(...)`are always called via reflection if the argument is `null` because the argument checking only allows for a non-null string. To allow argument checking to be a bit more forgiving, this change uses pattern-matching and removes the `ArgumentParser.TryGetArg(s)`.

Finally, overloads of `IntrinsicFunctions.NormalizePath` and `IntrinsicFunctions.NormlaizeDirectory` have been added that take a single path argument. This avoids an array being allocated every time these are called with a single string argument.