Move StringSyntaxAttribute polyfill to correct namespace
StringSyntaxAttribute should be defined in the System.Diagnostics.CodeAnalysis namespace but the MSBuild polyfill is in System. This change puts it in the right namespace and adds a type forward when on .NET.

In addition, Display the files from the Microsoft.CodeAnalysis.Contracts source package in a Contracts folder rather than the root of the project.